:bug: ワーカーを多重キルしてしまう問題の修正

+++ src/Assets/Scripts/GameMain/Task/PoisonArea.cs
@@ -121,7 +121,11 @@ namespace GameMain.Task
                     int index = Random.Range(0, workers.Count);
                     Worker worker = workers[index];
                     workers.RemoveAt(index);
-                    worker.Kill();
+
+                    if (!worker.IsLocked)
+                    {
+                        worker.Kill();
+                    }
                 }
 
                 await UniTask.Delay(TimeSpan.FromSeconds(workerDamageInterval), cancellationToken: cancellationToken);
